@@ -1,72 +1,71 @@
 {
-    "name": "date-time-picker-app",
-    "version": "0.0.0",
-    "description": "Angular Date Time Picker",
-    "keywords": [
-        "Angular",
-        "datepicker",
-        "date picker",
-        "timepicker",
-        "time picker"
-    ],
-    "author": "daniel pan <myron.yk.pan@gmail.com>",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build picker",
-        "test": "ng test picker",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "build_lib": "ng build picker && cp README.md dist/picker",
-        "build_css": "mkdir -p dist/picker/assets/style && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
-        "npm_pack": "cd dist/picker && npm pack",
-        "package": "npm run build_lib && npm run build_css && npm run npm_pack"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/DanielYKPan/date-time-picker.git"
-    },
-    "dependencies": {
-        "@angular/animations": "~7.2.14",
-        "@angular/common": "~7.2.14",
-        "@angular/compiler": "~7.2.14",
-        "@angular/core": "~7.2.14",
-        "@angular/forms": "~7.2.14",
-        "@angular/http": "~7.2.14",
-        "@angular/platform-browser": "~7.2.14",
-        "@angular/platform-browser-dynamic": "~7.2.14",
-        "@angular/router": "~7.2.14",
-        "core-js": "^2.6.5",
-        "moment": "^2.24.0",
-        "moment-timezone": "^0.5.25",
-        "rxjs": "~6.5.1",
-        "zone.js": "~0.9.0"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.8",
-        "@angular-devkit/build-ng-packagr": "~0.13.8",
-        "@angular/cdk": "^7.3.7",
-        "@angular/cli": "~7.3.8",
-        "@angular/compiler-cli": "~7.2.14",
-        "@angular/language-service": "~7.2.14",
-        "@types/jasmine": "~3.3.12",
-        "@types/jasminewd2": "~2.0.6",
-        "@types/node": "~11.13.7",
-        "codelyzer": "~5.0.0",
-        "jasmine-core": "~3.4.0",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.1.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.5",
-        "karma-jasmine": "~2.0.1",
-        "karma-jasmine-html-reporter": "^1.4.0",
-        "ng-packagr": "^5.1.0",
-        "protractor": "~5.4.2",
-        "ts-node": "~8.1.0",
-        "tsickle": ">=0.34.3",
-        "tslib": "^1.9.3",
-        "tslint": "~5.16.0",
-        "typescript": ">=3.1.1 <3.3.0"
-    }
+  "name": "date-time-picker-app",
+  "version": "0.0.0",
+  "description": "Angular Date Time Picker",
+  "keywords": [
+    "Angular",
+    "datepicker",
+    "date picker",
+    "timepicker",
+    "time picker"
+  ],
+  "author": "daniel pan <myron.yk.pan@gmail.com>",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build picker",
+    "test": "ng test picker",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "build_lib": "ng build picker && cp README.md dist/picker",
+    "build_css": "mkdir -p dist/picker/assets/style && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
+    "npm_pack": "cd dist/picker && npm pack",
+    "package": "npm run build_lib && npm run build_css && npm run npm_pack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/DanielYKPan/date-time-picker.git"
+  },
+  "dependencies": {
+    "@angular/animations": "~8.2.6",
+    "@angular/common": "~8.2.6",
+    "@angular/compiler": "~8.2.6",
+    "@angular/core": "~8.2.6",
+    "@angular/forms": "~8.2.6",
+    "@angular/platform-browser": "~8.2.6",
+    "@angular/platform-browser-dynamic": "~8.2.6",
+    "@angular/router": "~8.2.6",
+    "core-js": "^2.6.5",
+    "moment": "^2.24.0",
+    "moment-timezone": "^0.5.25",
+    "rxjs": "~6.5.3",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.4",
+    "@angular-devkit/build-ng-packagr": "~0.803.4",
+    "@angular/cdk": "^7.3.7",
+    "@angular/cli": "~8.3.4",
+    "@angular/compiler-cli": "~8.2.6",
+    "@angular/language-service": "~8.2.6",
+    "@types/jasmine": "~3.3.12",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~11.13.7",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^5.4.0",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.1.0",
+    "tsickle": "^0.37.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.16.0",
+    "typescript": "~3.5.3"
+  }
 }
