@@ -48,7 +48,7 @@ export const OWL_DATETIME_VALUE_ACCESSOR: any = {
 })
 export class OwlDateTimeInlineComponent<T> extends OwlDateTime<T>
     implements OnInit, ControlValueAccessor {
-    @ViewChild(OwlDateTimeContainerComponent)
+    @ViewChild(OwlDateTimeContainerComponent, { static: true })
     container: OwlDateTimeContainerComponent<T>;
 
     /**
